+++ src/main/java/net/aoba/gui/IMoveable.java
@@ -1,12 +1,12 @@
 package net.aoba.gui;
 
 public interface IMoveable {
-	public int getX();
-	public int getY();
-	public int getHeight();
-	public int getWidth();
-	public void setX(int x);
-	public void setY(int y);
-	public void setWidth(int width);
-	public void setHeight(int height);
+	public float getX();
+	public float getY();
+	public float getHeight();
+	public float getWidth();
+	public void setX(float x);
+	public void setY(float y);
+	public void setWidth(float width);
+	public void setHeight(float height);
 }
