+++ src/main/java/net/aoba/gui/hud/AbstractHud.java
@@ -14,10 +14,10 @@ import net.minecraft.client.gui.DrawContext;
 import java.util.function.Consumer;
 
 public abstract class AbstractHud implements IMoveable {
-	protected int x;
-	protected int y;
-	protected int width;
-	protected int height;
+	protected float x;
+	protected float y;
+	protected float width;
+	protected float height;
 	
 	protected RenderUtils renderUtils = new RenderUtils();
 	protected MinecraftClient mc = MinecraftClient.getInstance();
@@ -34,8 +34,8 @@ public abstract class AbstractHud implements IMoveable {
 		position_setting_update = new Consumer<Vector2>() {
 			@Override
 			public void accept(Vector2 vector2) {
-				setX((int)vector2.x);
-				setY((int)vector2.y);
+				setX(vector2.x);
+				setY(vector2.y);
 			}
 		};
 
@@ -43,37 +43,38 @@ public abstract class AbstractHud implements IMoveable {
 		SettingManager.register_setting(position_setting, Aoba.getInstance().settingManager.hidden_category);
 	}
 	
-	public int getX() {
+	public float getX() {
 		return this.x;
 	}
 	
-	public int getY() {
+	public float getY() {
 		return this.y;
 	}
 	
-	public int getHeight() {
+	public float getHeight() {
 		return this.height;
 	}
 	
-	public int getWidth() {
+	public float getWidth() {
 		return this.width;
 	}
 	
-	public void setX(int x) {
+	public void setX(float x) {
 		this.x = x;
 	}
 	
-	public void setY(int y) {
+	public void setY(float y) {
 		this.y = y;
 	}
 	
-	public void setWidth(int width) {
+	public void setWidth(float width) {
 		this.width = width;
 	}
 	
-	public void setHeight(int height) {
+	public void setHeight(float height) {
 		this.height = height;
 	}
+	
 	public void update(double mouseX, double mouseY, boolean mouseClicked) {
 		if (Aoba.getInstance().hudManager.isClickGuiOpen()) {
 			if (HudManager.currentGrabbed == null) {
